+++ attributes/default.rb
@@ -46,6 +46,8 @@ default['ntp']['listen_network'] = nil
 default['ntp']['apparmor_enabled'] = false
 default['ntp']['monitor'] = false
 default['ntp']['statistics'] = true
+default['ntp']['conf_restart_immediate'] = false
+
 # See http://www.vmware.com/vmtn/resources/238 p. 23 for explanation
 default['ntp']['disable_tinker_panic_on_virtualization_guest'] = true